Simulator Performance Upgrade
resolves #2 , Nested functions with global variables like
system_dynamics are now methods to eliminate the time it used to take
for MATLAB to search for the shared variables. Thruster combinations
that are used in calculating moment and forces are saved to expunge the
time of recalculating them. Interpolation method calls function
F_controller once instead of three time.