critical update on plot function, see description
add plots for the values of theta that are input to controllers, and the corresponding rotational speeds (from differentiating thetas). turns out, the rotational speeds are sometimes not aligned with the actual w1,w2,w3 states of the satellite. next commit must fix the inputs to the controller as (theta, thetadot) instead of w (omegas)